requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
pyyaml>=6.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
nltk>=3.7

# Machine learning and LLM tools
torch>=1.10.0
transformers>=4.15.0
datasets>=1.18.0
scikit-learn>=1.0.0
faiss-cpu>=1.7.0  # For semantic search in future updates
openai>=1.107.3
anthropic>=0.18.0  # topic_package LLM (optional)

# Visualization
plotly>=5.5.0
networkx>=2.6.0  # For attention flow visualizations

# Jupyter and reporting
jupyter>=1.0.0
nbformat>=5.1.0
nbconvert>=7.0.0
markdown>=3.3.0
ipython>=7.31.0
weasyprint>=56.0

# Terminal UI
rich>=13.0.0

# Utility packages
psutil>=5.9.0
requests>=2.27.0
python-dotenv>=0.19.0
colorama>=0.4.4
pypdf>=4.2.0

# Topic mining (citation_graph): embeddings + clustering
sentence-transformers>=2.2.0
hdbscan>=0.8.33
# Optional: leidenalg, python-igraph (for Leiden; otherwise Louvain is used)